🔧 Update deployed addresses and enhance automation scripts
✅ Updated `deployed-addresses-mock.json` with new token and collateral configurations to ensure accuracy in the system.
✅ Added new automation scripts for checking asset handler decimals and vault authorizations, improving diagnostics and error handling.
✅ Enhanced existing automation scripts for better clarity and functionality, including fixes for token decimal configurations.
✅ Updated Makefile with new commands for fixing automation configurations and generating correct checkData for Chainlink registration.
✅ Verified and tested all changes to ensure seamless integration and functionality within the overall system.

+++ script/automation/DiagnosePerformUpkeepFlow.s.sol
@@ -137,8 +137,8 @@ contract DiagnosePerformUpkeepFlow is Script {
         
         console.log("CheckData encoded for:");
         console.log("  Loan Manager:", automationAdapter);
-        console.log("  Start Index:", 0);
-        console.log("  Batch Size:", 25);
+        console.log("  Start Index:", uint256(0));
+        console.log("  Batch Size:", uint256(25));
         console.log("");
         
         try keeper.checkUpkeep(checkData) returns (bool upkeepNeeded, bytes memory performData) {
